+++ backend/src/main/java/com/factory/backend/services/IProductOrderService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface IProductOrderService {
     List<ProductOrderDTO> getAllProductOrders();
 
-    ProductOrderDTO getProductOrderById(Long productId, String clientPhone);
+    ProductOrderDTO getProductOrderById(Integer productId, String clientPhone);
 
     ProductOrderDTO saveProductOrder(ProductOrderAddingDTO ProductOrderDTO);
